@@ -1,22 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
         <IsPackable>false</IsPackable>
     </PropertyGroup>
     
     <ItemGroup>
-        <PackageReference Include="coverlet.msbuild" Version="2.5.1" />
-        <PackageReference Include="FluentAssertions" Version="5.6.0" />
-        <PackageReference Include="Microsoft.AspNetCore.All" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="2.2.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.2.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
+        <PackageReference Include="coverlet.msbuild" Version="2.8.0" />
+        <PackageReference Include="FluentAssertions" Version="5.10.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.1" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.1" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
         <PackageReference Include="NUnit" Version="3.11.0" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.13.0" />
-        <PackageReference Include="SonarAnalyzer.CSharp" Version="7.11.0.8083" />
-        <PackageReference Include="SpecFlow" Version="3.0.169-beta" />
-        <PackageReference Include="SpecFlow.NUnit" Version="3.0.169-beta" />
-        <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.0.169-beta" />
+        <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
+        <PackageReference Include="SonarAnalyzer.CSharp" Version="8.2.0.14119" />
+        <PackageReference Include="SpecFlow" Version="3.1.76" />
+        <PackageReference Include="SpecFlow.NUnit" Version="3.1.76" />
+        <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.1.76" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\Bookings.API\Bookings.API.csproj" />
