@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 namespace Bookings.API
 {
@@ -14,12 +14,13 @@ namespace Bookings.API
             CreateWebHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args)
+        private static IHostBuilder CreateWebHostBuilder(string[] args)
         {
-            return WebHost.CreateDefaultBuilder(args)
-                .UseKestrel(x => x.AddServerHeader = false)
-                .UseApplicationInsights()
-                .UseStartup<Startup>();
+            return Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder => { 
+                    webBuilder.UseStartup<Startup>();
+                    webBuilder.UseKestrel(x => x.AddServerHeader = false);
+                });
         }
     }
 }
