@@ -1,29 +1,22 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using Microsoft.AspNetCore.Mvc.Authorization;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace Bookings.API.Swagger
 {
     public class AuthResponsesOperationFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             var filterPipeline = context.ApiDescription.ActionDescriptor.FilterDescriptors;
             var isAuthorized = filterPipeline.Select(filterInfo => filterInfo.Filter)
                 .Any(filter => filter is AuthorizeFilter);
             var allowAnonymous = filterPipeline.Select(filterInfo => filterInfo.Filter)
                 .Any(filter => filter is IAllowAnonymousFilter);
-
+            
             if (!isAuthorized || allowAnonymous) return;
-
-            if (operation.Responses == null)
-            {
-                operation.Responses = new Dictionary<string, Response>();
-            }
-
-            operation.Responses.Add("401", new Response {Description = "Unauthorized"});
+            operation.Responses.Add("401", new OpenApiResponse { Description = "Unauthorized" });
         }
     }
 }
