@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Bookings.Api.Contract.Requests
+﻿namespace Bookings.Api.Contract.Requests
 {
     public interface IRepresentativeInfoRequest
     {
