@@ -1,44 +1,52 @@
-﻿namespace Bookings.Common
+﻿using System;
+using System.Collections.Generic;
+using System.Security.Principal;
+using Microsoft.ApplicationInsights;
+using Microsoft.ApplicationInsights.DataContracts;
+using Microsoft.ApplicationInsights.Extensibility;
+using Newtonsoft.Json;
+
+namespace Bookings.Common
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Security.Principal;
-    using Microsoft.ApplicationInsights;
-    using Microsoft.ApplicationInsights.DataContracts;
-    using Newtonsoft.Json;
 
     /// <summary>
     /// The application logger class send telemetry to Application Insights.
     /// </summary>
     public static class ApplicationLogger
     {
-        private static readonly TelemetryClient TelemetryClient = new TelemetryClient();
+        private static readonly TelemetryClient TelemetryClient = InitTelemetryClient();
+        
+        private static TelemetryClient InitTelemetryClient() {
+            var config = TelemetryConfiguration.CreateDefault();
+            var client = new TelemetryClient(config);
+            return client;
+        }
 
         public static void Trace(string traceCategory, string eventTitle, string information)
         {
-            var telematryTrace = new TraceTelemetry(traceCategory, severityLevel: SeverityLevel.Information);
-            telematryTrace.Properties.Add("Information", information);
-            telematryTrace.Properties.Add("Event", eventTitle);
-            TelemetryClient.TrackTrace(telematryTrace);
+            var traceTelemetry = new TraceTelemetry(traceCategory, severityLevel: SeverityLevel.Information);
+            traceTelemetry.Properties.Add("Information", information);
+            traceTelemetry.Properties.Add("Event", eventTitle);
+            TelemetryClient.TrackTrace(traceTelemetry);
         }
 
         public static void TraceWithProperties(string traceCategory, string eventTitle, string user, IDictionary<string, string> properties)
         {
-            var telematryTrace = new TraceTelemetry(traceCategory.ToString(), severityLevel: SeverityLevel.Information);
+            var traceTelemetry = new TraceTelemetry(traceCategory, SeverityLevel.Information);
 
-            telematryTrace.Properties.Add("Event", eventTitle);
+            traceTelemetry.Properties.Add("Event", eventTitle);
 
-            telematryTrace.Properties.Add("User", user);
+            traceTelemetry.Properties.Add("User", user);
 
             if (properties != null)
             {
                 foreach (KeyValuePair<string, string> entry in properties)
                 {
-                    telematryTrace.Properties.Add(entry.Key, entry.Value);
+                    traceTelemetry.Properties.Add(entry.Key, entry.Value);
                 }
             }
 
-            TelemetryClient.TrackTrace(telematryTrace);
+            TelemetryClient.TrackTrace(traceTelemetry);
           
         }
    
@@ -49,18 +57,18 @@
 
         public static void TraceWithObject(string traceCategory, string eventTitle, string user, object valueToSerialized)
         {
-            var telematryTrace = new TraceTelemetry(traceCategory.ToString(), severityLevel: SeverityLevel.Information);
+            var traceTelemetry = new TraceTelemetry(traceCategory, SeverityLevel.Information);
 
-            telematryTrace.Properties.Add("Event", eventTitle);
+            traceTelemetry.Properties.Add("Event", eventTitle);
 
-            telematryTrace.Properties.Add("User", user);
+            traceTelemetry.Properties.Add("User", user);
 
             if (valueToSerialized != null)
             {
-                telematryTrace.Properties.Add(valueToSerialized.GetType().Name, JsonConvert.SerializeObject(valueToSerialized, Formatting.None));
+                traceTelemetry.Properties.Add(valueToSerialized.GetType().Name, JsonConvert.SerializeObject(valueToSerialized, Formatting.None));
             }
 
-            TelemetryClient.TrackTrace(telematryTrace);
+            TelemetryClient.TrackTrace(traceTelemetry);
         }
 
         public static void TraceWithObject(string traceCategory, string eventTitle, string user)
