@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Bookings.Common.Helpers
+﻿namespace Bookings.Common.Helpers
 {
     public enum TraceCategory
     {
